requirements.txt
inquirerpy==0.3.4
pfzy==0.3.4
pyfiglet==1.0.2
PyJWT==2.8.0
readchar==4.0.5
requests==2.31.0
termcolor==2.3.0
tqdm==4.66.1
